pyproject.toml
[project]
name = "powerpoint-maker-ddb"
version = "0.1.0"
description = "A Python project to create PowerPoint presentations from text using OpenAI embeddings"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "python-pptx>=0.6.21",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "PyPDF2>=3.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
powerpoint-maker = "src.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

